+++ apps/clock_001/main.c
@@ -1,13 +1,3 @@
-/* Includes ------------------------------------------------------------------*/
-//#include "main.h"
-//#include "stm32f4xx_hal.h"
-//#include "adc.h"
-//#include "dma.h"
-//#include "spi.h"
-//#include "tim.h"
-//#include "usart.h"
-//#include "gpio.h"
-
 #include <stdint.h>
 #include <stdbool.h>
 #include "App.h"
