+++ apps/clock_001/makefile
@@ -1,14 +1,18 @@
 # STM32F4xx bsp uart test
 REPO_ROOT = ../..
 
+ifeq ($(strip $(board)),)
+ifneq ('$(MAKECMDGOALS)','clean')
+$(info no board specified, setting to clock)
+board = clock
+endif
+endif
+
 C_SOURCES =  \
 main.c \
 App.cpp \
 BlinkerPanel.cpp \
-MidiClockDisplay.cpp \
 midiTime.cpp \
-SegmentVideo.cpp \
-sketch.cpp \
 StatusPanel.cpp
 
 STARTUP_SOURCES =  \
@@ -23,13 +27,18 @@ $(REPO_ROOT)/bsp/CubeProject/Drivers/STM32F4xx_HAL_Driver/Inc \
 $(REPO_ROOT)/bsp/CubeProject/Drivers/CMSIS/Include \
 $(REPO_ROOT)/bsp/CubeProject/Drivers/CMSIS/Device/ST/STM32F4xx/Include \
 $(REPO_ROOT)/bsp/UserProject/AltDrivers \
-$(REPO_ROOT)/Arduino/Core \
-$(REPO_ROOT)/Arduino/Libs/uCModules/src \
-$(REPO_ROOT)/Arduino/Libs/arduino_midi/src
+$(REPO_ROOT)/modules/logging \
+$(REPO_ROOT)/modules/SegmentVideo \
+$(REPO_ROOT)/modules/midi47fx \
+$(REPO_ROOT)/modules/uCModules
+
 
 STATIC_LIBS = \
-$(REPO_ROOT)/Arduino/build/Arduino.a \
-$(REPO_ROOT)/bsp/build/bsp.a
+$(REPO_ROOT)/modules/SegmentVideo/build/SegmentVideo.a \
+$(REPO_ROOT)/modules/midi47fx/build/midi47fx.a \
+$(REPO_ROOT)/modules/uCModules/build/uCModules.a \
+$(REPO_ROOT)/modules/logging/build/logging.a \
+$(REPO_ROOT)/bsp/build/bsp_$(board).a
 
 #PROJECT_OBJECTS
 #OBJECTS = $(patsubst %s,%o,$(patsubst %c,%o,$(patsubst %cpp,%o,$(C_SOURCES))))
@@ -47,6 +56,9 @@ $(BUILD_DIR)/output.elf: build_dirs $(addprefix $(BUILD_DIR)/obj/, $(OBJECTS)) $
 	@echo "#### Linking into elf"
 	$(PP) -o $(TARGET_EXEC) $(LDFLAGS) -std=c++11 $(addprefix $(BUILD_DIR)/obj/, $(SOURCE_OBJECTS)) $(STATIC_LIBS) $(addprefix $(BUILD_DIR)/obj/, $(STARTUP_OBJECT))
 	$(SZ) $@
+	@echo ""
+	@echo "Built for $(board) board"
+	@echo ""
 	$(OBJD) -S --disassemble $(TARGET_EXEC) > $(TARGET_EXEC).dump
 	@echo
 
@@ -64,9 +76,7 @@ $(BUILD_DIR)/output.bin: build_dirs $(TARGET_EXEC)
 
 all: $(STATIC_LIBS) $(TARGET_EXEC)
 
-clean::
-	$(MAKE) -C $(REPO_ROOT)/bsp clean
-	$(MAKE) -C $(REPO_ROOT)/Arduino clean
+clean:: clean_all
 
 openocd:
 	cygstart cmd /k launch_openocd.bat
