--- apps/clock_001/sketch.h
@@ -1,12 +1,0 @@
-#ifndef SKETCH_H
-#define SKETCH_H
-
-#include "Arduino.h"
-#include "midiTime.h"
-
-#define Serial Serial6
-
-void setup(void);
-void loop(void);
-
-#endif
